requirements.txt
folium~=0.14.0
geohash2~=1.1
geopandas~=0.14.0
geopy~=2.4.0
hausdorff~=0.2.6
haversine~=2.8.0
idx2numpy~=1.2.3
ipython~=8.11.0
kaleido~=0.2.0
matplotlib~=3.7.2
notebook~=6.5.2
numba~=0.57.1
numpy~=1.24.3
opacus~=1.4.1
packaging~=23.2
pandas~=2.0.3
parameterized~=0.9.0
Pillow~=10.1.0
POT~=0.9.4
plotly~=5.15.0
scikit-learn~=1.3.0
scipy~=1.11.3
Shapely~=2.0.2
tabulate~=0.9.0
tensorboard~=2.16.2
tensorflow~=2.16.1
tensorrt~=8.6.1
torch~=2.1.2
torchaudio~=2.1.2
torchvision~=0.16.2
tqdm~=4.65.0
utm~=0.7.0
ipywidgets~=8.1.2